Now all tilemap targets are displayed in a vertical list, in order
The tilemap target dropdown has been hidden.
Selecting a target in the list selects the object in the scene.
Selecting a tilemap in the scene, selects the item in the list.
Removing or adding a tilemap in the scene should refresh the list.
The tileset grid has been stretched.
Tilemap layer items are sorted by type (which would normally match some Sorting layers) and by Sort order of each tilemap in every type group.
Created a settings asset where to define which are the standard layers for the project, used for sorting.