Added buttons to tilemap layers UI: creating new layers + moving layers
Now it is possible to create new tilemaps at the top of the layers of a type, or below all of them.
Now it is possible to change the sort order of a tilemap, which changes its position in the layer stack.
Fixed: Now it does not rebuilds the list every time a layer is selected.