Tile's layer selection + Tile's layer creation + Tile asset selection
Now when a tile is picked in the grid, the tilemap layer that matches the layer of the tile (layer type + sorting order) is automatically selected. There is a toggle that allows enabling / disabling this feature.
Added a button to the tileset grid that allows selecting the asset that corresponds to the current tile.
Added a button to create the tilemap layer of the current tile in case its layer does not exist.