Palette selection bar + Palette icons + Edit & DrawGizmos buttons moved
The palette dropdown has been removed. Instead there is a scroll view with buttons that correspond to all the palettes.
Added a button to the grid that allows storing the sprite of the current tile as the icon of the current palette, which is shown in its button.
Palette icons are stored in a new scriptable object.
The Edit and the Draw gizmos buttons have been moved to the grid.
Now the grid is highlighted when the Edit mode is on.
Fixed: The scroll view of the tilemap layer list was not appearing when moving the brush inspector area.