Now it works when there are no grids, no layers or no palettes + Fixes
When there is no grid in the scene, a text message is shown instead of the layer list.
When there are no layers, the layer type headers are shown anyway.
The height of the layer list has minimum limit of 5 rows (this avoids the brush inspector from doing strange things).
Fixed: The Create layer button was appearing when is should not. Now it is hidden when there is no grid or when the tile is not of the custom class.
Fixed: It was not possible to create a new layer when the grid was empty.
Now instantiated tilemaps are now linked to their prefabs.
Now new layers are automatically selected.
Fixed: Layer type headers were not being displayed when there were no layers of such types.